Created main file
Create thread 1 which runs face_rec() function. When the function start it will begin to detect the faces of individual that have been encoded in the EncodedFile.p. When q is pressed on the keyboard, the function will exit and will stop the program. This can be use to stop the whole program execution.