Progress 3
Successfully make the threads run together. The first problem encountered is when a new registration happened and the EncodedFile.p need to be encoded again, the threads will stop running due to file access conflict. The first approach is to use thread management system. However, the structure of the program itself, does not fit with the thread management. The successful approach is to disable file reading in the face_recognition function. This allow the EncodedFile.p to be able to encode again without interrupting the running threads.
Next task-> include GPS reading when individual face is detected.